LF-3592 update migration to fix sale_value to float

+++ packages/api/db/migration/20230901204048_create_general_sale_table.js
@@ -20,7 +20,7 @@
 export const up = async function (knex) {
   await knex.schema.createTable('general_sale', (table) => {
     table.integer('sale_id').references('sale_id').inTable('sale');
-    table.integer('sale_value');
+    table.float('sale_value');
     table.string('notes');
   });
 };
